pyproject.toml
[project]
name            = "wazuh-mcp-server"
version         = "0.1.0"
description     = "MCP tool-server for Wazuh Manager (stream transport)."
authors         = [{name = "Your Name", email = "you@example.com"}]
license         = {text = "Apache-2.0"}
readme          = "README.md"
requires-python = ">=3.11"

dependencies = [
  "pydantic>=2.7",
  "httpx[http2]>=0.27",
  "starlette>=0.37",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9",
  "mcp[cli]>=1.9.4",
  "sse-starlette",
  "PyYAML>=6.0",
  "fastmcp>=0.4"
]

[project.scripts]
wazuh-mcp-server = "wazuh_mcp_server.__main__:main"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]